refactor(rpc): split routes.rs into submodules (backlog #12 phase 1)
Move routes.rs → routes/mod.rs and split out three supporting
concerns into sibling submodules:

* routes/auth.rs — ApiKey extractor + constant_time_eq
* routes/ratelimit.rs — IpRateLimiter type, GlobalIpLimiter /
  WriteIpLimiter newtypes, extract_client_ip, check_rate_limit,
  rate_limit_response, the ip_rate_limit_middleware and
  write_rate_limit_middleware implementations
* routes/types.rs — ApiResponse / SendTxRequest / SignedTxRequest

Public API surface preserved via pub-use re-exports in mod.rs, so
existing call sites (jsonrpc/mod.rs, explorer*, lib.rs, tests,
main.rs) continue to resolve crate::routes::ApiKey / SharedState /
GlobalIpLimiter / etc. without any import churn.

Handlers (~900 lines of per-resource async fns) stay in mod.rs for
this phase. Phase 2 will split those by resource (chain / accounts
/ staking / tokens / epoch / ops).

mod.rs: 1593 → 1368 LOC. No behaviour change, clippy clean, all
existing tests pass.
